Final version
Some changes were made and a few bugs were fixed. Some notes,
the bug where after you win or lose in guess the number still isnt fixed.
Could not figure out how to get the int PlayerTurn function to work.
The tally add up in R,P,S has a bug in it and I still do not know how to fix it.